chore: move helper functions for bytes/equals to separate file
