+++ bytes/mod.ts
@@ -69,38 +69,6 @@ export function lastIndexOf(
   return -1;
 }
 
-function equalsNaive(a: Uint8Array, b: Uint8Array): boolean {
-  if (a.length !== b.length) return false;
-  for (let i = 0; i < b.length; i++) {
-    if (a[i] !== b[i]) return false;
-  }
-  return true;
-}
-
-function equalsSimd(a: Uint8Array, b: Uint8Array): boolean {
-  if (a.length !== b.length) return false;
-  const len = a.length;
-  const compressable = Math.floor(len / 4);
-  const compressedA = new Uint32Array(a.buffer, 0, compressable);
-  const compressedB = new Uint32Array(b.buffer, 0, compressable);
-  for (let i = compressable * 4; i < len; i++) {
-    if (a[i] !== b[i]) return false;
-  }
-  for (let i = 0; i < compressedA.length; i++) {
-    if (compressedA[i] !== compressedB[i]) return false;
-  }
-  return true;
-}
-
-/** Check whether binary arrays are equal to each other.
- * @param a first array to check equality
- * @param b second array to check equality
- */
-export function equals(a: Uint8Array, b: Uint8Array): boolean {
-  if (a.length < 1000) return equalsNaive(a, b);
-  return equalsSimd(a, b);
-}
-
 /** Check whether binary array starts with prefix.
  * @param source source array
  * @param prefix prefix array to check in source
@@ -209,3 +177,5 @@ export function copy(src: Uint8Array, dst: Uint8Array, off = 0): number {
 
 /** @deprecated */
 export { includes as contains };
+
+export { equals } from "./equals.ts";
